rename: 통계 계산 기능 파일 이름 재설정
